CBG-715 Attach CompareHashAndPassword to auth cache
